requirements.txt
azure-ai-projects
azure-identity